Code developed to answer peer review of associated paper.
This code was written to generate modified versions of the original figures in the main paper. Instead of weighting GFMPT uniformly, the "weighted" versions of each code will weight each MFPT by the stationary distribution of the two nodes being traveled between. The "edge series time analysis computes TA-GFMPT for a list of video frames of a network and plots it. "Critical edge map" will compute and plot critical edges of a network as red and all others as black. Node_Stationary_Dist will create a color plot of the stationary distribution of a network.